--- lizrd/archive/run_train.sh
@@ -1,11 +1,0 @@
-#!/bin/bash
-#
-#SBATCH --job-name=lizard_train_sparse6
-#SBATCH --partition=common
-#SBATCH --qos=24gpu7d
-#SBATCH --gres=gpu:titanv:1
-#SBATCH --time=0-04:00:00
-#SBATCH --output=/home/jaszczur/sbatchlogs_s6.txt
-
-source venv/bin/activate
-python3 -m lizrd.train.bert_train
