+++ lizrd/core/misc.py
@@ -5,7 +5,6 @@ from torch.utils.checkpoint import checkpoint
 
 import torch.nn as nn
 from lizrd.core.initialization import get_init_weight
-from lizrd.support import ash
 
 
 class Noop(nn.Module):
@@ -66,7 +65,7 @@ class EinMix(nn.Module):
         return newoutput
 
 
-@ash.check("... inp -> ... out")
+
 def DenseEinMix(dinp, dout):
     return EinMix(
         "... dinp -> ... dout",
@@ -77,7 +76,7 @@ def DenseEinMix(dinp, dout):
     )
 
 
-@ash.check("... inp -> ... out")
+
 class Linear(nn.Linear):
     def __init__(self, *args, init_type, init_scale, **kwargs):
         if "bias" not in kwargs:
@@ -102,7 +101,7 @@ def check_layer_funs(*layer_funs):
             )
 
 
-@ash.check("... -> ...")
+
 class StopGradient(nn.Module):
     def __init__(self):
         super(StopGradient, self).__init__()
@@ -115,7 +114,7 @@ def stop_gradient(x):
     return x.detach()
 
 
-@ash.check("... -> ...")
+
 class StopValuePassGradient(nn.Module):
     def __init__(self):
         super(StopValuePassGradient, self).__init__()
