+++ research/conditional/moe_layers/_expert_choice_old.py
@@ -1,4 +1,5 @@
 from typing import Literal, Union, Optional
+import einops
 import plotly.express as px
 import torch
 import torch.nn.functional as F
@@ -7,7 +8,6 @@ from torch.nn import LayerNorm
 
 import torch.nn as nn
 from lizrd.core.initialization import get_init_weight
-from lizrd.support import ash
 from lizrd.support.logging import make_histogram
 from lizrd.train import checkpointing
 from research.conditional.utils.layer_manager import LoggingLayer
@@ -474,7 +474,7 @@ class ExpertChoiceFFOld(LoggingLayer):
                 x,
                 self.lin2_weight,
             )
-            ash.assert_shape("e k m", x, e=self.n_experts, k=topk, m=self.doutput)
+            self.assertEqual(x.shape, (self.n_experts, topk, self.doutput))
         return x
 
     def gating_postprocess_onehot(
@@ -497,7 +497,7 @@ class ExpertChoiceFFOld(LoggingLayer):
     ):
         # multiply by softmax
         with measure_time(self, "multiply_softmax"):
-            ash.assert_shape("e k", topk_values, e=self.n_experts, k=topk)
+            self.assertEqual(topk_values.shape, (self.n_experts, topk))
             x = einsum(
                 "n_exp topk dmodel, n_exp topk -> n_exp topk dmodel", x, topk_values
             )
