+++ research/conditional/moe_layers/cont_moe_designs/add_layernorms.py
@@ -3,11 +3,10 @@ import torch
 from lizrd.core import misc
 import torch.nn as nn
 import lizrd.core.initialization
-from lizrd.support import ash
 from research.conditional.moe_layers.continuous_moe import ContinuousMoeBaseClass
 
 
-@ash.check("... dinp -> ... dout")
+
 class ContinuousMoELayernorm(ContinuousMoeBaseClass):
     def merge_map_emit(self, x, merge_weights, emit_weights):
         x = misc.einsum(
