+++ research/conditional/moe_layers/cont_moe_designs/no_softmax_on_weights.py
@@ -3,11 +3,9 @@ import torch
 from lizrd.core import misc
 import torch.nn as nn
 import lizrd.core.initialization
-from lizrd.support import ash
 from research.conditional.moe_layers.continuous_moe import ContinuousMoeBaseClass
 
 
-@ash.check("... dinp -> ... dout")
 class ContinuousMoENosoftmax(ContinuousMoeBaseClass):
     """
     The merging and emitting is done with just a linear layer, no softmax.
