+++ research/conditional/moe_layers/cont_moe_designs/send_result_only_to_top1_token.py
@@ -1,11 +1,27 @@
+import torch
 from lizrd.core import misc
-from lizrd.support import ash
-from research.conditional.archive.rogue_code import set_highest_index_one
 from research.conditional.moe_layers.continuous_moe import ContinuousMoeBaseClass
 from research.conditional.utils.misc_tools import stable_softmax_temperature
 
 
-@ash.check("... dinp -> ... dout")
+def set_highest_index_one(tensor: torch.Tensor) -> torch.Tensor:
+    # Get the index of the highest value in the last dimension
+    _, indices = torch.max(tensor, dim=-1)
+
+    # Create a new tensor filled with zeros, with the same shape as the input tensor
+    result_tensor = torch.zeros_like(tensor)
+
+    # Calculate index shape for the new tensor
+    result_shape = list(range(tensor.dim() - 1))
+
+    # Set 1 at the index of the highest value in each sub-array
+    result_tensor.scatter_(
+        -1, indices.view(tensor.shape[:-1] + (1,)).expand(tensor.shape), 1
+    )
+
+    return result_tensor
+
+
 class ContinuousMoETopmerge(ContinuousMoeBaseClass):
     """
     The emit only sends the output to the token that had the highest logit in the group.
